<UPDATE> Experimental changes for improvements
1. Replace "Aliveness checking" with GatewayReqs to mitigates the
half-duplex problem and its impact.

2. Add a field of nextReqTime in the GatewayReq to inform the node
the approximate time when the next GatewayReq is expected to arrive.
Thus if the node does not receive the request in the expected time
period, it will think that a path is no longer possible between itself
and the gateway. Then it goes on a self-healing process.

3. Some minor changes to make the code cleaner.