<UPDATE> Use broadcasts for request
Change to using broadcast for gateway request. Request that is not from the parent node will be filtered out in the forwarding engine.

The reason for using broadcast is to make transmissions more efficient. Therefore, it further reduces the possibility of the half-duplex problem and collisions between child nodes and parent nodes.