<FIX> Fix an issue due to receiving timeout
Apply the same receving timeout when the program starts to receive a specific number of bytes in the MessageProcessor.

The previous implementation only applies the receiving timeout at a message level. Problem rises when a gateway receives a corrupted message and blocks to receive bytes that do not exist.  This can cause the parent node to wait for an arbitarily long time period and lead to false-positive self-healing in its children nodes