<UPDATE> Adaptive backoff time
Adaptive backoff time enables a parent node or gateway to inform its child nodes of the maximum backoff time based on the number of child nodes it has currently connected to.

Current implementation gives 3 seconds of maximum backoff time for 1 child. For example,

1 child node -> 3 seconds
2 child nodes -> 6 seconds
 ...

We still applies backoff time even if the parent node has 1 child node. There is a uncertain case where two far-away parent nodes each have a child node and those child nodes are in range. Thus, if the requests from two parent nodes arrive simultaneously (but no collision since two parent nodes are far away), the child nodes can avoid collision with the backoff time. But, we need more knowledge on if this case exists.

Adaptive backoff time is not used for JoinACK messages which have the chance of collision dependent on the number of nearby connected nodes instead of the number of child nodes.