<UPDATE> Add comments for ATmega32u4 boards
Added some comments as notes to the user.

The while loops in the examples for waiting the serial port to open can cause program to hang for ATmega32u4-based boards. So it is better to comment them out if you don't want to connect the board to a USB port for debugging.